Split market request models
Move each Market request model into its own file.

Use long for Market ID fields while keeping documented string execution IDs unchanged.

Update changelog and tests for the model layout/type corrections.

+++ Bybit.Api/Market/Responses/BybitMarketInstrument.cs
@@ -8,7 +8,7 @@ public record BybitMarketSpotInstrument
     /// <summary>
     /// The ID of symbol name
     /// </summary>
-    public int? SymbolId { get; set; }
+    public long? SymbolId { get; set; }
 
     /// <summary>
     /// Symbol name
@@ -176,7 +176,7 @@ public record BybitMarketFuturesInstrument
     /// <summary>
     /// The ID of symbol name
     /// </summary>
-    public int? SymbolId { get; set; }
+    public long? SymbolId { get; set; }
 
     /// <summary>
     /// Symbol name
@@ -509,7 +509,7 @@ public record BybitMarketOptionInstrument
     /// <summary>
     /// The ID of symbol name
     /// </summary>
-    public int? SymbolId { get; set; }
+    public long? SymbolId { get; set; }
 
     /// <summary>
     /// Symbol name
